+++ server/Definitions/SystemStats.cs
@@ -5,6 +5,9 @@ namespace Napack.Server
     /// <summary>
     /// Holds statistics for the Napack Framework Server system as a whole.
     /// </summary>
+    /// <remarks>
+    /// TODO this needs a DB redesign as this isn't cheap / will be updated for every single request!!
+    /// </remarks>
     public class SystemStats
     {
         public SystemStats()
@@ -15,6 +18,10 @@ namespace Napack.Server
         public RequestStats RequestStats { get; set; }
 
         public Dictionary<string, string> IpToCountry { get; set; }
+
+        public int TotalNapacksDownloaded { get; set; }
+
+        public int TotalNapacksViewed { get; set; }
 
         public string LookupCountry(string ip)
         {